==> WebScraping/code/01_dexters_scrapper.ipynb <==
from google.colab import drive
drive.mount('/content/drive/')

import sys
from importlib import reload
reload(sys)


import pandas as pd
import requests
from joblib import Parallel, delayed             
import pickle                                   
from tqdm import tqdm                            


def html_extractor(url):                         
    try:
        cookies = dict(cookies_are='working')
        r = requests.get(url, cookies=cookies)
        return r.text
    except:
        return "no html"


clickbaits = pd.read_csv('/content/drive/My Drive/web_scrap/data/Clickbait.csv')

non_clickbaits = pd.read_csv('/content/drive/My Drive/web_scrap/data/NonClickbait.csv')

clickbait_urls = clickbaits.status_link.values                    
non_clickbait_urls = non_clickbaits.status_link.values


clickbait_html = Parallel(n_jobs=40)(delayed(html_extractor)(u) for u in tqdm(clickbait_urls))    
pickle.dump(clickbait_html, open('/content/drive/My Drive/web_scrap/data/clickbait_html.pkl', 'wb'), -1)

non_clickbait_html = Parallel(n_jobs=40)(delayed(html_extractor)(u) for u in tqdm(non_clickbait_urls))
pickle.dump(non_clickbait_html, open('/content/drive/My Drive/web_scrap/data/non_clickbait_html.pkl', 'wb'), -1)
