==> WebScraping/code/05_dexters_LSTM.ipynb <==

from google.colab import drive
drive.mount('/content/drive/')


#Import all the required libraries
import pandas as pd
from keras.models import Sequential
from keras.layers import Dense, Activation, Flatten
from keras.layers.embeddings import Embedding
from keras.layers.recurrent import LSTM
from keras.layers.normalization import BatchNormalization
from keras.utils import np_utils
from keras.callbacks import ModelCheckpoint
from keras.layers.advanced_activations import PReLU
from keras.preprocessing import sequence, text
from keras.layers import SpatialDropout1D
from keras import metrics

#Load the training and test data
train = pd.read_csv('/content/drive/My Drive/web_scrap/data/train.csv')
test = pd.read_csv('/content/drive/My Drive/web_scrap/data/test.csv')

y_train = train.label.values
y_test = test.label.values

tk = text.Tokenizer(num_words=200000)
train.link_name = train.link_name.astype(str)
test.link_name = test.link_name.astype(str)
train.textdata = train.textdata.astype(str)
test.textdata = test.textdata.astype(str)

max_len = 80

tk.fit_on_texts(list(train.link_name.values) + list(train.textdata.values) + list(test.link_name.values) + list(
    test.textdata.values))
x_train_title = tk.texts_to_sequences(train.link_name.values)
x_train_title = sequence.pad_sequences(x_train_title, maxlen=max_len)

x_train_textdata_01 = tk.texts_to_sequences(train.textdata.values)
x_train_textdata_01 = sequence.pad_sequences(x_train_textdata_01, maxlen=max_len)

x_test_title_01 = tk.texts_to_sequences(test.link_name.values)
x_test_title_01 = sequence.pad_sequences(x_test_title_01, maxlen=max_len)

x_test_textdata_02 = tk.texts_to_sequences(test.textdata.values)
x_test_textdata_02 = sequence.pad_sequences(x_test_textdata_02, maxlen=max_len)

word_index = tk.word_index
ytrain_enc = np_utils.to_categorical(y_train)

lstm_check_model = Sequential()
lstm_check_model.add(Embedding(len(word_index) + 1, 300, input_length=80, dropout=0.2),)
lstm_check_model.add(LSTM(300, return_sequences=True, dropout=0.2, recurrent_dropout=0.2))

lstm_check_model.add(Dense(200))
lstm_check_model.add(PReLU())
lstm_check_model.add(SpatialDropout1D(0.2))

lstm_check_model.add(BatchNormalization())

lstm_check_model.add(Dense(200))
lstm_check_model.add(PReLU())
lstm_check_model.add(SpatialDropout1D(0.2))
lstm_check_model.add(BatchNormalization())

lstm_check_model.add(Flatten())

lstm_check_model.add(Dense(2))
lstm_check_model.add(Activation('softmax'))


lstm_check_model.compile(loss='categorical_crossentropy', optimizer='adam', metrics=['acc' ])


checkpoint = ModelCheckpoint('/content/drive/My Drive/web_scrap/data/weights.h5', monitor='val_acc', save_best_only=True, verbose=2)

lstm_check_model.fit(x_train_title, y=ytrain_enc,
                 batch_size=128, epochs=20, verbose=2, validation_split=0.1,
                 shuffle=True, callbacks=[checkpoint])
